Rename server/finder to server/find
