+++ server/find/handler/cached_stats.go
