+++ server/find/http/handler_test.go
@@ -1,4 +1,4 @@
-package httpfinderserver
+package httpfindserver
 
 import (
 	"bytes"
@@ -15,7 +15,7 @@ import (
 	"github.com/ipfs/go-cid"
 	"github.com/ipni/go-indexer-core"
 	"github.com/ipni/go-libipni/find/model"
-	"github.com/ipni/storetheindex/server/finder/test"
+	"github.com/ipni/storetheindex/server/find/test"
 	"github.com/ipni/storetheindex/test/util"
 	"github.com/libp2p/go-libp2p/core/peer"
 	"github.com/multiformats/go-multiaddr"
