+++ server/find/http/header_test.go
@@ -1,4 +1,4 @@
-package httpfinderserver
+package httpfindserver
 
 import (
 	"net/http"
