+++ server/find/libp2p/handler.go
@@ -19,7 +19,7 @@ import (
 	"github.com/ipni/storetheindex/internal/libp2pserver"
 	"github.com/ipni/storetheindex/internal/metrics"
 	"github.com/ipni/storetheindex/internal/registry"
-	"github.com/ipni/storetheindex/server/finder/handler"
+	"github.com/ipni/storetheindex/server/find/handler"
 	"github.com/libp2p/go-libp2p/core/peer"
 	"github.com/libp2p/go-libp2p/core/protocol"
 	"go.opencensus.io/stats"
