+++ server/find/libp2p/protocol_test.go
@@ -9,8 +9,8 @@ import (
 	p2pclient "github.com/ipni/go-libipni/find/client/p2p"
 	"github.com/ipni/storetheindex/internal/counter"
 	"github.com/ipni/storetheindex/internal/registry"
-	p2pserver "github.com/ipni/storetheindex/server/finder/libp2p"
-	"github.com/ipni/storetheindex/server/finder/test"
+	p2pserver "github.com/ipni/storetheindex/server/find/libp2p"
+	"github.com/ipni/storetheindex/server/find/test"
 	"github.com/libp2p/go-libp2p"
 	"github.com/libp2p/go-libp2p/core/host"
 	"github.com/libp2p/go-libp2p/core/peer"
