+++ server/find/test/test_extended_providers.go
